tpdd filesize only 16bit, fix bootstrap post-install, ioctl()->fcntl(), baudrates

+++ dir_list.h
@@ -30,7 +30,7 @@ typedef struct
 {
 	char     client_fname[TPDD_FILENAME_LEN+1];
 	char     local_fname[LOCAL_FILENAME_MAX];
-	uint32_t len;
+	uint16_t len;
 	uint8_t  flags;
 } FILE_ENTRY;
